v0.5.4: Deep source audit fixes, OrangePi build+benchmark verification
Fix critical xcorr OOB reads in NEON/SVE2 (4 functions), moderate MTI
underflow, Vulkan init retry leak, vkMapMemory error checking, NEON sin
UB, SPIR-V alignment, build-time SPIR-V target and WSI crash fixes.
All 16 tests pass, 5 benchmark suites verified on Orange Pi 6 Plus.

+++ src/neon/neon_kernels.cpp
@@ -511,7 +511,7 @@ void neon_fast_sin_f32(float* out, const float* in, std::size_t n) {
         p = c1 + p * x2;
         p = p * x;
 
-        if ((int)k & 1) p = -p;
+        if ((int64_t)k & 1) p = -p;
         out[i] = p;
     }
 #else
